rename null-telemetry-client > console-telemetry-client (filenames)
