+++ src/background/telemetry/console-telemetry-client.ts
