+++ src/tests/unit/tests/background/telemetry/console-telemetry-client.test.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import { ConsoleTelemetryClient } from 'background/telemetry/null-telemetry-client';
+import { ConsoleTelemetryClient } from 'background/telemetry/console-telemetry-client';
 import { TelemetryBaseData } from 'background/telemetry/telemetry-base-data';
 import { TelemetryLogger } from 'background/telemetry/telemetry-logger';
 import { IMock, Mock, Times } from 'typemoq';
